Move preset helpers to backend
