+++ tests/test_core.py
@@ -5,6 +5,7 @@ import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 import core
+from backend import presets
 from backend import metrics, exercises
 from backend.exercise import Exercise
 from backend.workout_session import WorkoutSession
@@ -70,9 +71,9 @@ def sample_db(tmp_path):
 
 
 def test_load_workout_presets(sample_db):
-    presets = core.load_workout_presets(sample_db)
-    assert core.WORKOUT_PRESETS == presets
-    assert presets == [
+    loaded = presets.load_workout_presets(sample_db)
+    assert presets.WORKOUT_PRESETS == loaded
+    assert loaded == [
         {
             "name": "Push Day",
             "exercises": [
@@ -278,7 +279,7 @@ def test_delete_metric_type_in_use_by_preset_exercise(sample_db):
 
 
 def test_find_presets_and_apply_changes(sample_db):
-    names = core.find_presets_using_exercise("Push Up", db_path=sample_db)
+    names = presets.find_presets_using_exercise("Push Up", db_path=sample_db)
     assert names == ["Push Day"]
 
     ex = Exercise("Push Up", db_path=sample_db, is_user_created=False)
@@ -306,7 +307,7 @@ def test_find_presets_and_apply_changes(sample_db):
     }
     conn.close()
 
-    core.apply_exercise_changes_to_presets(ex, ["Push Day"], db_path=sample_db)
+    presets.apply_exercise_changes_to_presets(ex, ["Push Day"], db_path=sample_db)
 
     conn = sqlite3.connect(sample_db)
     cur = conn.cursor()
