+++ ui/screens/edit_preset_screen.py
@@ -26,9 +26,9 @@ from kivymd.uix.dialog import MDDialog
 import os
 import sqlite3
 
-import core
 from backend import metrics, exercises
-from core import DEFAULT_DB_PATH
+from backend.presets import load_workout_presets
+from core import DEFAULT_DB_PATH, DEFAULT_SETS_PER_EXERCISE
 
 
 class SectionWidget(MDBoxLayout):
@@ -335,7 +335,7 @@ class EditPresetScreen(MDScreen):
                 new_exercises.append(
                     {
                         "name": ex.get("name"),
-                        "sets": ex.get("sets") or core.DEFAULT_SETS_PER_EXERCISE,
+                        "sets": ex.get("sets") or DEFAULT_SETS_PER_EXERCISE,
                         "rest": ex.get("rest") or session.rest_duration,
                         "results": results,
                         "library_exercise_id": ex.get("library_id"),
@@ -613,7 +613,7 @@ class EditPresetScreen(MDScreen):
         def do_confirm(*args):
             try:
                 app.preset_editor.save()
-                core.load_workout_presets(app.preset_editor.db_path)
+                load_workout_presets(app.preset_editor.db_path)
                 app.selected_preset = app.preset_editor.preset_name
                 if dialog:
                     dialog.dismiss()
